Make RollupDriverBuilder implementation context-generic
