+++ crates/sovereign/sovereign-cosmos-relayer/src/contexts/mod.rs
@@ -1,4 +1,5 @@
 pub mod cosmos_to_sovereign_relay;
 pub mod sovereign_chain;
 pub mod sovereign_cosmos_birelay;
+pub mod sovereign_rollup;
 pub mod sovereign_to_cosmos_relay;
