+++ crates/sovereign/sovereign-cosmos-relayer/src/contexts/sovereign_rollup.rs
@@ -1,0 +1,1 @@
+pub struct SovereignRollup {}
