+++ crates/sovereign/sovereign-integration-tests/src/impls/mod.rs
@@ -1,0 +1,1 @@
+pub mod build_rollup_driver;
