+++ crates/sovereign/sovereign-integration-tests/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod contexts;
+pub mod impls;
